Revert "update: wait running apt processes"
This reverts commit 31470473f8418a47156545d73e4952ef7edbcc48.

+++ build-devel.sh
@@ -17,14 +17,7 @@
 # this is a replica of ncp-update, but from devel branch
 install() 
 {
-  # wait for other apt processes
-  test -f /var/lib/apt/lists/lock && echo "waiting for apt processes to finish..." && \
-  while :; do
-    test -f /var/lib/apt/lists/lock && break
-    sleep 1
-  done
-  rm -f /etc/apt/apt.conf.d/20nextcloudpi-upgrades
-
+  test -f /etc/apt/apt.conf.d/20nextcloudpi-upgrades && { echo "deactivate unattended upgrades first"; return 1; }
   echo -e "Downloading updates"
   rm -rf /tmp/ncp-update-tmp
   git clone --depth 20 -q -b devel https://github.com/nextcloud/nextcloudpi.git /tmp/ncp-update-tmp
